refactor: extract fetching resources to custom hook
